Separate relays.rs into multiple modules
